+++ js/util.js
@@ -1,0 +1,18 @@
+const getRandInt = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandArrElem = (elements) => elements[getRandInt(0, elements.length - 1)];
+
+const getId = () => {
+  let lastGeneratedId = 0;
+  return function () {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+export {getRandInt, getRandArrElem, getId};
